feat: extend qpscs to 8

+++ Source/Lib/Codec/pcs.h
@@ -1262,7 +1262,7 @@ typedef struct PictureControlSetInitData {
     uint8_t variance_boost_strength;
     uint8_t variance_octile;
     uint8_t tf_strength;
-    uint8_t qp_scale_compress_strength;
+    double  qp_scale_compress_strength;
     bool    max_32_tx_size;
     bool    adaptive_film_grain;
     uint8_t noise_norm_strength;
